Fixed type hash calculations tu ise 64-bit values

+++ include/memoria/v1/allocators/inmem/threads/allocator.hpp
@@ -1263,10 +1263,10 @@ private:
         int32_t page_size;
         in >> page_size;
 
-        int32_t ctr_hash;
+        uint64_t ctr_hash;
         in >> ctr_hash;
 
-        int32_t page_hash;
+        uint64_t page_hash;
         in >> page_hash;
 
         unique_ptr<int8_t, void (*)(void*)> page_data((int8_t*)::malloc(page_data_size), ::free);
